requirements.txt
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2
natsort

numpy==1.25.0
cython_bbox
cython
opencv-python
scikit-image
scikit-learn
matplotlib
tqdm
Pillow==9.5.0
ninja

pycocotools
yacs
easydict
sympy==1.13.1
filelock
typing-extensions>=4.10.0
fsspec
networkx
jinja2
